backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pinecone-client==3.0.0
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.6
python-multipart==0.0.6
